update 1.2, check README

+++ main.py
@@ -3,46 +3,69 @@ import configparser as cf # Для парса конфига
 import os, json, shutil, sys
 from pprint import pprint
 from progress.bar import IncrementalBar
-from ya_metods import * # Там код для работы с диском
+from metods.ya_metods import * # Там код для работы с диском
+from metods.vk_metods import * # Там код для работы с вк
 
 config = cf.ConfigParser()
 config.read("config.ini")
-disk_token = config.get('TOKEN','disk_token') # Токен диска
 serv_key = config.get('TOKEN','serv_key') # Сервисный ключ доступа 🤖
 
-# Проверим введение данных в конфиге
-if disk_token == 'your token':
-    print('Вы не ввели токен в конфиг!')
-    sys.exit()
-elif serv_key == 'your key':
-    print('Вы не ввели ключ в конфиг!')
-    sys.exit()
+if serv_key == 'your key':
+    print('Вставьте свой сервисный ключ ВК АПИ в файл config.ini!')
 
+from_album = input('Из какого альбома будем получать фотографии?(wall - фото со стены, profile - аватарка профиля, a - отдельное id фльбома.)\nНапишите w/p/a :|')
+if from_album not in ['w', 'p', 'a', 'W', 'P', 'A']:
+    print('Не верная команда!')
+    sys.exit()
+disk_token = input('Введите ваш токен от Яндекс Диска: |')
 own_id = input('Введите id пользователя ВКонтакте: |')
+fold_name = input('Название создаваемой папки на диске: |')
 
 # Очистим папку
-try:
-    os.makedirs('saved_pictures/piks_out')
-except FileExistsError:
-    print('')
+try: os.makedirs('saved_pictures/piks_out')
+except FileExistsError: print('')
 shutil.rmtree('saved_pictures/piks_out')
 os.makedirs('saved_pictures/piks_out')
 
-bar = IncrementalBar('Скачивание с ВК:', max = 6) 
-
-bar.next()
-URL = 'https://api.vk.com/method/photos.get' 
-params = {
-    'access_token': serv_key,
-    'owner_id':own_id,
-    'album_id':'profile',
-    'v':'5.131',
-    'extended':'1'
-}
+bar = IncrementalBar('Скачивание с ВК:', max = 5) 
 
 bar.next()
 
-res = requests.get(URL, params=params) # Пускаем запрос 🚀
+res = VKontakte() # Пускаем запрос 🚀
+alb_cheker = VKontakte()
+
+if from_album == ('p' or 'P'):
+    from_what = 'profile'
+    res = res.downlphoto(serv_key, own_id, from_what)
+elif from_album == ('w' or 'W'):
+    from_what = 'wall'
+    res = res.downlphoto(serv_key, own_id, from_what)
+elif from_album == ('a' or 'A'):
+    res = res.getalbums(serv_key, own_id)
+    print(f"\n\nВыберите альбом: ")
+    for i in res.json()['response']['items']:
+        print(f"\nНомер альбома: {i['id']}\nНазвание альбома: {i['title']}")
+    album_id = input("\nВведите номер нужного альбома: |")
+    from_what = album_id
+    res = alb_cheker.downlphoto(serv_key, own_id, from_what)
+
+if 'error' in res.json():
+    print('\n\nОшибка!\n')
+    if res.json()['error']['error_code'] == 100:
+        print('Запрашиваемый id пользователя не найден!')
+        sys.exit()
+    elif res.json()['error']['error_code'] == 300:
+        print('Альбом переполнен! Перед продолжением работы нужно удалить лишние объекты из альбома или использовать другой альбом.')
+    elif res.json()['error']['error_code'] == 30:
+        print('Профиль является приватным!')
+    elif res.json()['error']['error_code'] == 29:
+        print('Программа перегружена :(\nПопробуйте ещё раз через некоторое время.')
+    elif res.json()['error']['error_code'] == 18:
+        print('Страница пользователя была удалена или заблокирована.')
+    elif res.json()['error']['error_code'] == 10:
+        print('Произошла внутренняя ошибка сервера. Попробуйте позже!')
+    elif res.json()['error']['error_code'] == 6:
+        print('Слишком много запросой в секунду. Попробуйте позже!')
 
 bar.next()
 
@@ -78,11 +101,11 @@ with open('saved_pictures/saved_pics.json', 'w', encoding='utf-8') as json_file:
 bar.next()
 bar.finish()
 
-# Полетели на диск
+# Полетели на диск 💽
 
-bar = IncrementalBar('Загрузка на диск:', max = 3 + len(new_json)*2)
+bar = IncrementalBar('Загрузка на диск:', max = 2 + len(new_json)*2)
 ya = YandexDisk(disk_token)
-ya.create_folder('VK photos') # создаём папку для выгрузки
+ya.create_folder(fold_name) # создаём папку для выгрузки 📂
 
 bar.next()
 
@@ -96,12 +119,13 @@ bar.next()
 
 for image in new_json: # Отправим на диск 👾
     try:
-        ya.upload_photo(f'VK photos/{image}', f'saved_pictures/piks_out/{image}.jpg')
+        ya.upload_photo(f'{fold_name}/{image}', f'saved_pictures/piks_out/{image}.jpg')
         bar.next()
     except requests.exceptions.HTTPError:
         print('\nФото не найдено! Возможно оно не существует. Проверьте ссылку', new_json.get(image))
         bar.next()
         continue
+
 bar.next()
 bar.finish()
